# Reduce comparisons when parsing numbers