@headius 2024-11-07
Seems like it would be an easy port to the Java parser, if anyone gets to it before I do.